reorganize package structure and add project type
