+++ code/iaas/api-logic/src/main/java/io/cattle/platform/iaas/api/auth/github/resource/Project.java
@@ -1,0 +1,9 @@
+package io.cattle.platform.iaas.api.auth.github.resource;
+
+import io.cattle.platform.core.model.tables.records.AccountRecord;
+import io.github.ibuildthecloud.gdapi.annotation.Type;
+
+@Type(name="project")
+public class Project extends AccountRecord {
+
+}
